# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /valuation/{id}:
    post:
      operationId: runValuation
      tags: [valuation]
      summary: Run an AI valuation of an analysis's finish/repurpose repos
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Portfolio valuation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioValuation"
    get:
      operationId: getValuation
      tags: [valuation]
      summary: Get a previously computed portfolio valuation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Portfolio valuation, or null if not yet computed
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/PortfolioValuation"
                  - type: "null"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ValuationFactor:
      type: object
      required: [label, score, weight, detail]
      properties:
        label:
          type: string
        score:
          type: number
        weight:
          type: number
        detail:
          type: string
    ValuationRevenuePotential:
      type: object
      required: [model, monthly_revenue_low, monthly_revenue_high, timeline]
      properties:
        model:
          type: string
        monthly_revenue_low:
          type: integer
        monthly_revenue_high:
          type: integer
        timeline:
          type: string
    ValuationComparable:
      type: object
      required: [name, outcome, multiple, relevance]
      properties:
        name:
          type: string
        outcome:
          type: string
        multiple:
          type: string
        relevance:
          type: string
    Valuation:
      type: object
      required:
        [
          repo, estimated_value_low, estimated_value_high, currency, valuation_method,
          confidence, factors, revenue_potential, comparables, risks, upsides, summary,
        ]
      properties:
        repo:
          type: string
        estimated_value_low:
          type: integer
        estimated_value_high:
          type: integer
        currency:
          type: string
        valuation_method:
          type: string
        confidence:
          type: string
          enum: [low, medium, high]
        factors:
          type: array
          items:
            $ref: "#/components/schemas/ValuationFactor"
        revenue_potential:
          $ref: "#/components/schemas/ValuationRevenuePotential"
        comparables:
          type: array
          items:
            $ref: "#/components/schemas/ValuationComparable"
        risks:
          type: array
          items:
            type: string
        upsides:
          type: array
          items:
            type: string
        summary:
          type: string
    PortfolioValuationTopPick:
      type: object
      required: [repo, reason]
      properties:
        repo:
          type: string
        reason:
          type: string
    PortfolioValuation:
      type: object
      required:
        [
          total_estimated_value_low, total_estimated_value_high, currency, repo_valuations,
          portfolio_summary, top_picks, diversification_score, recommendation,
        ]
      properties:
        total_estimated_value_low:
          type: integer
        total_estimated_value_high:
          type: integer
        currency:
          type: string
        repo_valuations:
          type: array
          items:
            $ref: "#/components/schemas/Valuation"
        portfolio_summary:
          type: string
        top_picks:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioValuationTopPick"
        diversification_score:
          type: number
        recommendation:
          type: string
...